+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -8,6 +8,6 @@ public interface UserStorage {
     User addUser(User user);
     User putUser(User user);
     List<User> getUsers();
-    User findUser(Integer id);
-    User deleteUser(Integer id);
+    User findUser(Long id);
+    User deleteUser(Long id);
 }
